BL-2999 make a separate class to handle current book
(That is, to handle server requests that depend on knowing the
current book.)
