+++ src/BloomExe/ProjectContext.cs
@@ -223,6 +223,7 @@ namespace Bloom
 //				{
 					_httpServer = new EnhancedImageServer(new RuntimeImageProcessor(bookRenameEvent), parentContainer.Resolve<BookThumbNailer>());
 					ReadersHandler.Init(_httpServer);
+					CurrentBookHandler.Init(_httpServer);
 //				}
 					builder.Register((c => _httpServer)).AsSelf().SingleInstance();
